+++ packages/amplify-ui-components/src/global/theme.ts
@@ -54,8 +54,8 @@ if (browserOrNode().isBrowser) {
       --amplify-light-grey: #c4c4c4;
       --amplify-white: #ffffff;
       --amplify-smoke-white: #f5f5f5;
-
       --amplify-red: #dd3f5b;
+      --amplify-blue: #099ac8;
     }
   `),
   );