+++ packages/api-graphql/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.1](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-graphql@1.2.0...@aws-amplify/api-graphql@1.2.1) (2020-09-03)
+
+**Note:** Version bump only for package @aws-amplify/api-graphql
+
+
+
+
+
+# [1.2.0](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-graphql@1.1.7...@aws-amplify/api-graphql@1.2.0) (2020-09-03)
+
+
+### Features
+
+* **SSR:** withSSRContext ([#6146](https://github.com/aws-amplify/amplify-js/issues/6146)) ([1cb1afd](https://github.com/aws-amplify/amplify-js/commit/1cb1afd1e56135908dceb2ef6403f0b3e78067fe))
+
+
+
+
+
+## [1.1.7](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-graphql@1.1.6...@aws-amplify/api-graphql@1.1.7) (2020-09-01)
+
+**Note:** Version bump only for package @aws-amplify/api-graphql
+
+
+
+
+
 ## [1.1.6](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-graphql@1.1.5...@aws-amplify/api-graphql@1.1.6) (2020-08-19)
 
 **Note:** Version bump only for package @aws-amplify/api-graphql